fix+perf: audit fixes, security hardening, and performance optimizations
Security fixes:
- Eliminate SQL injection in roles/workplaces dependency & nuke handlers
- Fix 13 auth.profile_id → acting_user_id bugs in marketplace (generic accounts)
- Add auth + can_access_diary permission check to diary GET endpoint
- Fix raw body parsing in get_locum_users (use Json<T> extractor)
- Add target shift ownership validation in perform_shift_swap
- Replace unwrap() panics with proper error handling (marketplace dates, shifts, diary)
- Add month/year range validation for marketplace queries
- Fix time format normalization in templates (prevent double :00)
- Fix COD role_id type from i64 to i32

Performance optimizations:
- Increase connection pool from 5 to 25 with min 2 idle connections
- Parallelize 9 dependency COUNT queries with tokio::try_join! (~9x faster)
- Add moka profile cache in auth extractor (60s TTL, eliminates DB hit per request)
- Add permission check caching (30s TTL, lazy static)
- Add reference data caching for workplaces and roles (60s TTL, auto-invalidated)
- Add 14 database indexes on frequently queried columns
- Add emoji prefixes to all tracing log messages

+++ src/models/comment.rs
@@ -6,7 +6,7 @@ use utoipa::ToSchema;
 #[derive(Debug, Clone, Serialize, Deserialize, FromRow, ToSchema)]
 pub struct COD {
     pub id: i64,  // INT8 in database (bigserial)
-    pub role_id: i64,  // INT8 in database
+    pub role_id: i32,  // INT4 - references Roles.id
     pub date: NaiveDate,
     pub created_by: i32,  // INT4 - references Users.user_profile_id
     pub comment: Option<String>,
